Reorg lambda functions & fix bug for getFeedList
