+++ lib/constants/apiLambdaConst.ts
@@ -1,3 +1,3 @@
-export const LAMBDA_FUNCTION_NAME = "MainHandler";
+export const GET_FEED_LIST_FUNCTION_NAME = "GetFeedListHandler";
 
 export const TEST_API_KEY_NAME = "test-api-key";
